[refactor] Move get_min_max_timestamps() and get_inventory_timestamps() to alaqsdblite.py; address 3 TODOs in code to simplify it.
